Fix some bug base on test_simple

+++ src/core/util/collection_util.rs
@@ -26,7 +26,7 @@ impl CollectionUtil {
     // TODO: does this suitable for Rust HashMap?
     /// Currently not used in Rust Lucene
     #[allow(dead_code)]
-    pub fn new_hashmap<K, V>(size: i32) -> HashMap<K, V> {
+    pub fn new_hashmap<K, V>(size: usize) -> HashMap<K, V> {
         let capacity = ((size as f32) / 0.75f32 + 1f32) as usize;
         HashMap::with_capacity(capacity)
     }
